+++ reader/src/main/java/dev/runo/reader/domain/repository/ReaderRepository.kt
@@ -4,6 +4,6 @@ import dev.runo.core.common.WorkStatus
 import dev.runo.reader.domain.model.Chapter
 
 interface ReaderRepository {
-    suspend fun getAllChapters(titleId: Long): WorkStatus<List<Chapter>>
+    suspend fun getAllChapters(titleId: Int): WorkStatus<List<Chapter>>
     suspend fun getImagesByChapterId(chapterId: Int): WorkStatus<List<String>>
 }
